requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
django-filter
drf-yasg
django-rest-swagger
djangorestframework-simplejwt
django-cors-headers